plastic-id-ml/pyproject.toml
# ───────────────────────────────────────────────────────────────
[tool.poetry]
name        = "plastic-id-ml"
version     = "0.1.0"
description = "ML classification of polymers - M.Sc. thesis"
authors     = ["Anders Blomberg <s339872@oslomet.no>"]

packages = [{ include = "plastic_id", from = "src" }]

# ───────────────────────────────────────────────────────────────
[tool.poetry.dependencies]          # — runtime, light
python       = "^3.10"
pandas       = "^2.2"
scikit-learn = "1.3.2"
hydra-core   = "^1.3"
typer        = "^0.12"
rich         = "^13.7"

# ───────────────────────────────────────────────────────────────
numpy = "1.26.4"
[tool.poetry.group.ml.dependencies] # — heavy ML libs
xgboost = "~2.0"
torch   = "^2.3.1"

[tool.poetry.group.dev.dependencies]   # — dev / thesis tooling
pytest      = "^8.1"
joblib      = "^1.4.2"
matplotlib  = "^3.10"      # plotting lives only here now

# ───────────────────────────────────────────────────────────────
tabulate = "^0.9.0"
torch = "^2.7.0"
torchvision = "^0.22.0"
[tool.poetry.extras]                # — optional pip install .[viz]
viz = ["matplotlib"]

# ───────────────────────────────────────────────────────────────
[tool.poetry.scripts]
pid = "plastic_id.cli:app"
# ───────────────────────────────────────────────────────────────
